+++ bfabric/src/bfabric/examples/compare_zeep_suds_query.py
@@ -52,7 +52,7 @@ def full_query(auth: BfabricAuth, query: dict, includedeletableupdateable: bool 
     thisQuery = deepcopy(query)
     thisQuery["includedeletableupdateable"] = includedeletableupdateable
 
-    return {"login": auth.login, "password": auth.password, "query": thisQuery}
+    return {"login": auth.login, "password": auth.password.get_secret_value(), "query": thisQuery}
 
 
 def calc_both(