+++ bfabric_app_runner/src/bfabric_app_runner/cli/chunk.py
@@ -1,28 +1,23 @@
-from __future__ import annotations
-
 from pathlib import Path
 
-import cyclopts
-
+from bfabric import Bfabric
+from bfabric.experimental.entity_lookup_cache import EntityLookupCache
+from bfabric.experimental.workunit_definition import WorkunitDefinition
+from bfabric.utils.cli_integration import use_client
 from bfabric_app_runner.app_runner.resolve_app import load_workunit_information
 from bfabric_app_runner.app_runner.runner import run_app, Runner
 from bfabric_app_runner.output_registration import register_outputs
-from bfabric import Bfabric
-from bfabric.utils.cli_integration import setup_script_logging
-from bfabric.experimental.entity_lookup_cache import EntityLookupCache
-from bfabric.experimental.workunit_definition import WorkunitDefinition
-
-app_chunk = cyclopts.App("chunk", help="Run an app on a chunk. You can create the chunks with `app dispatch`.")
 
 
-@app_chunk.command()
-def run_all(
+@use_client
+def cmd_chunk_run_all(
     app_spec: Path,
     work_dir: Path,
     workunit_ref: int | Path,
     *,
     ssh_user: str | None = None,
     read_only: bool = False,
+    client: Bfabric,
 ) -> None:
     """Run all chunks, including input preparation, processing, and output registration.
 
@@ -32,9 +27,6 @@ def run_all(
     :param ssh_user: SSH user to use for downloading the input files, instead of the current user.
     :param read_only: If True, results will not be registered and the workunit status will not be changed.
     """
-    setup_script_logging()
-    client = Bfabric.from_config()
-
     app_version, workunit_ref = load_workunit_information(
         app_spec=app_spec, client=client, work_dir=work_dir, workunit_ref=workunit_ref
     )
@@ -50,16 +42,14 @@ def run_all(
     )
 
 
-@app_chunk.command()
-def process(app_spec: Path, chunk_dir: Path) -> None:
+@use_client
+def cmd_chunk_process(app_spec: Path, chunk_dir: Path, *, client: Bfabric) -> None:
     """Process a chunk.
 
     Note that the input files must be prepared before running this command.
     :param app_spec: Path to the app spec file.
     :param chunk_dir: Path to the chunk directory.
     """
-    setup_script_logging()
-    client = Bfabric.from_config()
     chunk_dir = chunk_dir.resolve()
 
     # TODO this lookup of workunit_definition is very problematic now! FIX NEEDED
@@ -72,8 +62,8 @@ def process(app_spec: Path, chunk_dir: Path) -> None:
         runner.run_process(chunk_dir=chunk_dir)
 
 
-@app_chunk.command()
-def outputs(
+@use_client
+def cmd_chunk_outputs(
     app_spec: Path,
     chunk_dir: Path,
     workunit_ref: int | Path,
@@ -82,6 +72,7 @@ def outputs(
     force_storage: Path | None = None,
     read_only: bool = False,
     reuse_default_resource: bool = True,
+    client: Bfabric,
 ) -> None:
     """Register the output files of a chunk.
 
@@ -92,9 +83,7 @@ def outputs(
     :param read_only: If True, the workunit will not be set to processing.
     :param reuse_default_resource: If True, the default resource will be reused for the output files. (recommended)
     """
-    # TODO redundant with "outputs register"
-    setup_script_logging()
-    client = Bfabric.from_config()
+    # TODO redundant functionality with "outputs register"
     chunk_dir = chunk_dir.resolve()
 
     app_version, workunit_ref = load_workunit_information(