+++ bfabric_scripts/src/bfabric_scripts/bfabric_logthis.py
@@ -9,7 +9,7 @@ import argparse
 from typing import Literal
 
 from bfabric import Bfabric
-from bfabric_scripts.cli.api.cli_api_log import write_externaljob, write_workunit
+from bfabric_scripts.cli.api._deprecated_log import write_externaljob, write_workunit
 from bfabric.utils.cli_integration import use_client