+++ bfabric_scripts/src/bfabric_scripts/cli/cli_executable.py
@@ -1,8 +1,8 @@
 import cyclopts
 
-from bfabric_scripts.cli.executable.inspect import inspect_executable
-from bfabric_scripts.cli.executable.upload import upload_executable
+from bfabric_scripts.cli.executable.show import cmd_executable_show
+from bfabric_scripts.cli.executable.upload import cmd_executable_upload
 
-app = cyclopts.App()
-app.command(inspect_executable, name="inspect")
-app.command(upload_executable, name="upload")
+cmd_executable = cyclopts.App(help="Read and write executable entities in B-Fabric.")
+cmd_executable.command(cmd_executable_show, name="show")
+cmd_executable.command(cmd_executable_upload, name="upload")