+++ bfabric_scripts/src/bfabric_scripts/cli/cli_workunit.py
@@ -1,10 +1,10 @@
 import cyclopts
 
-from bfabric_scripts.cli.workunit.export_definition import export_definition
+from bfabric_scripts.cli.workunit.export_definition import cmd_workunit_export_definition
 from bfabric_scripts.cli.workunit.not_available import (
-    list_not_available_proteomics_workunits,
+    cmd_workunit_not_available,
 )
 
-app = cyclopts.App()
-app.command(list_not_available_proteomics_workunits, name="not-available")
-app.command(export_definition, name="export-definition")
+cmd_workunit = cyclopts.App(help="Read workunit entities in B-Fabric.")
+cmd_workunit.command(cmd_workunit_not_available, name="not-available")
+cmd_workunit.command(cmd_workunit_export_definition, name="export-definition")