+++ bfabric_scripts/docs/changelog.md
@@ -10,6 +10,15 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[Unreleased\]
 
+## \[1.13.23\] - 2025-02-19
+
+### Added
+
+- `bfabric-cli dataset {upload, download, show}` to replace the old dataset-related scripts.
+- `bfabric-cli api update` command to update an existing entity
+- `bfabric-cli api create` command to create a new entity
+- `bfabric-cli api delete` command to delete an existing entity
+
 ## \[1.13.22\] - 2025-02-17
 
 ### Fixed